+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### BREAKING CHANGES
+- Convert package type to "module".
+- Dropped support for Node.js 14.
+
+### Removed
+- Support for eslint-plugin-jest in this package, linting rules for test files will be a different package.
 
 ## [10.0.0] - 2023-05-22
 ### BREAKING CHANGES